grpc test refactor (kind of)
